refactor: centralize floating buttons component
