+++ src/Reports/allOrder.jsx
@@ -3,11 +3,6 @@ import axios from "axios";
 import Skeleton, { SkeletonTheme } from "react-loading-skeleton";
 import "react-loading-skeleton/dist/skeleton.css";
 import { useNavigate } from "react-router-dom";
-import enquiry from '../assets/enquiry.svg';
-import payment from '../assets/payment.svg';
-import reciept from '../assets/reciept.svg';
-import FloatingButtons from "../Pages/floatingButton";
-import order from '../assets/order.svg';
 
 const AddOrder1 = lazy(() => import("../Pages/addOrder1"));
 const OrderUpdate = lazy(() => import("./orderUpdate"));
@@ -113,13 +108,6 @@ export default function AllOrder() {
         setSelectedOrder(null);
     };
 
-    const buttonsList = [
-        { onClick: () => navigate('/addTransaction'), src: reciept },
-        { onClick: () => navigate('/addTransaction1'), src: payment },
-        { onClick: () => navigate('/addOrder1'), src: order },
-        { onClick: () => navigate('/addEnquiry'), src: enquiry },
-    ];
-
     return (
         <>
             <div className="order-update-content bg-[#e5ddd5] min-h-screen">
@@ -216,12 +204,6 @@ export default function AllOrder() {
                         </div>
                     </main>
 
-                    <FloatingButtons
-                        buttonType="bars"
-                        buttonsList={buttonsList}
-                        direction="up"
-                        buttonClassName="rounded-full bg-green-500 hover:bg-green-600 text-white"
-                    />
                 </div>
 
                 <Suspense fallback={
